Pushed algorithm workaround for final API implementation
We actually did not need to change the weighting of the algorithm. Instead, all we had to do was change the distances we record while making a graph to be close to 0 if the location contains "Tunnel" in its name. This isn't best practices, but is a remarkably easy workaround for a project that could take hours to test.